Adding some useful tools our our common workflows
