+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "bootstrap": "^4.5.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.3"
@@ -30,5 +31,18 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
+    "jest-environment-jsdom-fourteen": "^1.0.1",
+    "jest-enzyme": "^7.1.2",
+    "jest-resolve": "^26.5.2",
+    "jest-useragent-mock": "^0.1.1",
+    "jest-watch-typeahead": "^0.6.1",
+    "node-sass": "^4.14.1",
+    "popper.js": "^1.16.1",
+    "sass-lint": "^1.13.1",
+    "typescript": "^4.0.3"
   }
 }
