+++ src/containers/App/App.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { mount } from 'enzyme';
+
+import App from './App';
+
+it('mounts without crashing', () => {
+  const component = mount(<App />);
+
+  expect(component).toMatchSnapshot();
+});
