+++ src/index.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+
+import App from './containers/App/App';
 import * as serviceWorker from './serviceWorker';
+
+import './sass/index.scss';
 
 ReactDOM.render(
   <React.StrictMode>
